Move icons components into icons folder
