+++ src/components/MappingEvents/Statistics.js
@@ -4,8 +4,8 @@ import React from 'react';
 import { t } from 'ttag';
 import styled from 'styled-components';
 
-import MapPinWithPlusIcon from './MapPinWithPlusIcon';
-import BellIcon from './BellIcon';
+import MapPinWithPlusIcon from '../icons/ui-elements/MapPinWithPlusIcon';
+import BellIcon from '../icons/ui-elements/BellIcon';
 import { UserIcon } from '../icons/ui-elements/index';
 
 type Props = {
