+++ src/components/icons/ui-elements/MapPinWithPlusIcon.js
